Replace unsigned int with size_t
Replace unsigned int definitions with size_t if they
describing the size of an object.

+++ rtrlib/spki/spkitable.h
@@ -108,7 +108,7 @@ int spki_table_add_entry(struct spki_table *spki_table, struct spki_record *spki
  * @return SPKI_SUCCESS On success
  * @return SPKI_ERROR On error
  */
-int spki_table_get_all(struct spki_table *spki_table, uint32_t asn, uint8_t *ski, struct spki_record **result, unsigned int *result_size);
+int spki_table_get_all(struct spki_table *spki_table, uint32_t asn, uint8_t *ski, struct spki_record **result, size_t *result_size);
 
 /**
  * @brief Returns all spki_record whose SKI number matches the given one.
@@ -119,7 +119,7 @@ int spki_table_get_all(struct spki_table *spki_table, uint32_t asn, uint8_t *ski
  * @return SPKI_SUCCESS On success
  * @return SPKI_ERROR On error
  */
- int spki_table_search_by_ski(struct spki_table *spki_table, uint8_t *ski, struct spki_record **result, unsigned int *result_size);
+ int spki_table_search_by_ski(struct spki_table *spki_table, uint8_t *ski, struct spki_record **result, size_t *result_size);
 
 
 /**
